+++ src/fbv/resource.h
@@ -10,6 +10,14 @@
 #define IDS_STRING108                   108
 #define IDS_COM_ERROR_FORMAT            108
 #define IDS_SHELL_VALIDATE_VERB        109
+#define IDS_STATUS_DONE                110
+#define IDS_BUTTON_STOP                111
+#define IDS_BUTTON_REVALIDATE          112
+#define IDS_COLUMN_FILE_NAME           113
+#define IDS_VALIDATION_NO_ERRORS       114
+#define IDS_SAX_ERROR_LOCATION         115
+#define IDS_BUTTON_EXIT                116
+#define IDS_STATUS_SCANNING            117
 #define IDC_FILELIST                    1000
 #define IDC_MSG                         1001
 #define IDC_VALIDATE                    1002
@@ -19,7 +27,7 @@
 // 
 #ifdef APSTUDIO_INVOKED
 #ifndef APSTUDIO_READONLY_SYMBOLS
-#define _APS_NEXT_RESOURCE_VALUE        107
+#define _APS_NEXT_RESOURCE_VALUE        116
 #define _APS_NEXT_COMMAND_VALUE         40001
 #define _APS_NEXT_CONTROL_VALUE         1005
 #define _APS_NEXT_SYMED_VALUE           101